+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.1.0] - UNRELEASED
 
 - Refactor namespaces to avoid conflict with UnityEditor.Editor class
+- Update Project Settings and Preferences path to CS Build Management
 
 ## [1.0.0] - 2020-08-08